refactor: rename folder component/styles/pageStyled.ts to component/styles/componentStyled.ts
