+++ components/styles/componentStyled.ts
